+++ FleetJourney.Tests/Mapping/EmployeeMapperTests.cs
@@ -5,7 +5,7 @@ using FleetJourney.Domain.EmployeeInfo;
 using FluentAssertions;
 using Xunit;
 
-namespace FleetJourney.Application.Tests.Mapping;
+namespace FleetJourney.Tests.Mapping;
 
 public sealed class EmployeeMapperTests
 {
