+++ FleetJourney.Tests/Validation/CarPool/CarResponseValidatorTests.cs
@@ -1,110 +1,94 @@
-﻿using FleetJourney.Application.Contracts.Requests.CarPool;
+﻿using FleetJourney.Application.Contracts.Responses.CarPool;
 using FleetJourney.Application.Validation.CarPool;
 using FluentValidation.TestHelper;
 using Xunit;
 
-namespace FleetJourney.Application.Tests.Validation.CarPool;
+namespace FleetJourney.Tests.Validation.CarPool;
 
-public sealed class CreateCarRequestValidatorTests
+public sealed class CarResponseValidatorTests
 {
-    private readonly CreateCarRequestValidator _validator;
+    private readonly CarResponseValidator _validator;
 
-    public CreateCarRequestValidatorTests()
+    public CarResponseValidatorTests()
     {
-        _validator = new CreateCarRequestValidator();
+        _validator = new CarResponseValidator();
     }
 
     [Theory]
-    [InlineData(null)]
     [InlineData("")]
+    [InlineData(null)]
     public async Task SetLicensePlateNumber_WithInvalidLicensePlateNumber_ShouldReturnValidationError(string licensePlateNumber)
     {
         // Arrange
-        var createCarRequest = new CreateCarRequest
+        var carResponse = new CarResponse
         {
+            Id = Guid.NewGuid(),
             LicensePlateNumber = licensePlateNumber,
             Brand = "Valid Brand",
             Model = "Valid Model",
             EndOfLifeMileage = 10000,
             MaintenanceInterval = 5000,
-            CurrentMileage = 5000
+            CurrentMileage = 5000,
+            Trips = null
         };
 
         // Act
-        var result = await _validator.TestValidateAsync(createCarRequest);
+        var result = await _validator.TestValidateAsync(carResponse);
 
         // Assert
-        result.ShouldHaveValidationErrorFor(request => request.LicensePlateNumber)
+        result.ShouldHaveValidationErrorFor(response => response.LicensePlateNumber)
             .WithErrorMessage("License plate number is required.");
     }
 
-    [Fact]
-    public async Task SetLicensePlateNumber_WithInvalidFormat_ShouldReturnValidationError()
-    {
-        // Arrange
-        var createCarRequest = new CreateCarRequest
-        {
-            LicensePlateNumber = "InvalidFormat",
-            Brand = "Valid Brand",
-            Model = "Valid Model",
-            EndOfLifeMileage = 10000,
-            MaintenanceInterval = 5000,
-            CurrentMileage = 5000
-        };
-
-        // Act
-        var result = await _validator.TestValidateAsync(createCarRequest);
-
-        // Assert
-        result.ShouldHaveValidationErrorFor(request => request.LicensePlateNumber)
-            .WithErrorMessage("Invalid license plate number. Format should be 123-AB-45.");
-    }
-
     [Theory]
-    [InlineData(null)]
     [InlineData("")]
+    [InlineData(null)]
     public async Task SetBrand_WithInvalidBrand_ShouldReturnValidationError(string brand)
     {
         // Arrange
-        var createCarRequest = new CreateCarRequest
+        var carResponse = new CarResponse
         {
+            Id = Guid.NewGuid(),
             LicensePlateNumber = "Valid License Plate Number",
             Brand = brand,
             Model = "Valid Model",
             EndOfLifeMileage = 10000,
             MaintenanceInterval = 5000,
-            CurrentMileage = 5000
+            CurrentMileage = 5000,
+            Trips = null
         };
 
         // Act
-        var result = await _validator.TestValidateAsync(createCarRequest);
+        var result = await _validator.TestValidateAsync(carResponse);
 
         // Assert
-        result.ShouldHaveValidationErrorFor(request => request.Brand)
+        result.ShouldHaveValidationErrorFor(response => response.Brand)
             .WithErrorMessage("Brand is required.");
     }
 
     [Theory]
-    [InlineData(null)]
     [InlineData("")]
+    [InlineData(null)]
     public async Task SetModel_WithInvalidModel_ShouldReturnValidationError(string model)
     {
         // Arrange
-        var createCarRequest = new CreateCarRequest
+        var carResponse = new CarResponse
         {
+            Id = Guid.NewGuid(),
             LicensePlateNumber = "Valid License Plate Number",
             Brand = "Valid Brand",
             Model = model,
             EndOfLifeMileage = 10000,
             MaintenanceInterval = 5000,
-            CurrentMileage = 5000
+            CurrentMileage = 5000,
+            Trips = null
         };
 
         // Act
-        var result = await _validator.TestValidateAsync(createCarRequest);
+        var result = await _validator.TestValidateAsync(carResponse);
 
         // Assert
-        result.ShouldHaveValidationErrorFor(request => request.Model)
+        result.ShouldHaveValidationErrorFor(response => response.Model)
             .WithErrorMessage("Model is required.");
     }
 
@@ -113,21 +97,23 @@ public sealed class CreateCarRequestValidatorTests
     public async Task SetEndOfLifeMileage_WithInvalidEndOfLifeMileage_ShouldReturnValidationError(uint endOfLifeMileage)
     {
         // Arrange
-        var createCarRequest = new CreateCarRequest
+        var carResponse = new CarResponse
         {
+            Id = Guid.NewGuid(),
             LicensePlateNumber = "Valid License Plate Number",
             Brand = "Valid Brand",
             Model = "Valid Model",
             EndOfLifeMileage = endOfLifeMileage,
             MaintenanceInterval = 5000,
-            CurrentMileage = 5000
+            CurrentMileage = 5000,
+            Trips = null
         };
 
         // Act
-        var result = await _validator.TestValidateAsync(createCarRequest);
+        var result = await _validator.TestValidateAsync(carResponse);
 
         // Assert
-        result.ShouldHaveValidationErrorFor(request => request.EndOfLifeMileage)
+        result.ShouldHaveValidationErrorFor(response => response.EndOfLifeMileage)
             .WithErrorMessage("End of life mileage must be greater than zero.");
     }
 
@@ -136,21 +122,23 @@ public sealed class CreateCarRequestValidatorTests
     public async Task SetMaintenanceInterval_WithInvalidMaintenanceInterval_ShouldReturnValidationError(uint maintenanceInterval)
     {
         // Arrange
-        var createCarRequest = new CreateCarRequest
+        var carResponse = new CarResponse
         {
+            Id = Guid.NewGuid(),
             LicensePlateNumber = "Valid License Plate Number",
             Brand = "Valid Brand",
             Model = "Valid Model",
             EndOfLifeMileage = 10000,
             MaintenanceInterval = maintenanceInterval,
-            CurrentMileage = 5000
+            CurrentMileage = 5000,
+            Trips = null
         };
 
         // Act
-        var result = await _validator.TestValidateAsync(createCarRequest);
+        var result = await _validator.TestValidateAsync(carResponse);
 
         // Assert
-        result.ShouldHaveValidationErrorFor(request => request.MaintenanceInterval)
+        result.ShouldHaveValidationErrorFor(response => response.MaintenanceInterval)
             .WithErrorMessage("Maintenance interval must be greater than zero.");
     }
 
@@ -159,21 +147,47 @@ public sealed class CreateCarRequestValidatorTests
     public async Task SetCurrentMileage_WithInvalidCurrentMileage_ShouldReturnValidationError(uint currentMileage)
     {
         // Arrange
-        var createCarRequest = new CreateCarRequest
+        var carResponse = new CarResponse
         {
+            Id = Guid.NewGuid(),
             LicensePlateNumber = "Valid License Plate Number",
             Brand = "Valid Brand",
             Model = "Valid Model",
             EndOfLifeMileage = 10000,
             MaintenanceInterval = 5000,
-            CurrentMileage = currentMileage
+            CurrentMileage = currentMileage,
+            Trips = null
         };
 
         // Act
-        var result = await _validator.TestValidateAsync(createCarRequest);
+        var result = await _validator.TestValidateAsync(carResponse);
 
         // Assert
-        result.ShouldHaveValidationErrorFor(request => request.CurrentMileage)
+        result.ShouldHaveValidationErrorFor(response => response.CurrentMileage)
             .WithErrorMessage("Current mileage must be greater than zero.");
+    }
+
+    [Fact]
+    public async Task SetId_WithEmptyId_ShouldReturnValidationError()
+    {
+        // Arrange
+        var carResponse = new CarResponse
+        {
+            Id = Guid.Empty,
+            LicensePlateNumber = "Valid License Plate Number",
+            Brand = "Valid Brand",
+            Model = "Valid Model",
+            EndOfLifeMileage = 10000,
+            MaintenanceInterval = 5000,
+            CurrentMileage = 5000,
+            Trips = null
+        };
+
+        // Act
+        var result = await _validator.TestValidateAsync(carResponse);
+
+        // Assert
+        result.ShouldHaveValidationErrorFor(response => response.Id)
+            .WithErrorMessage("Id is required.");
     }
 }
