+++ FleetJourney.Tests/Validation/CarPool/UpdateCarRequestValidatorTests.cs
@@ -1,94 +1,110 @@
-﻿using FleetJourney.Application.Validation.CarPool;
-using FleetJourney.Application.Contracts.Responses.CarPool;
+﻿using FleetJourney.Application.Contracts.Requests.CarPool;
+using FleetJourney.Application.Validation.CarPool;
 using FluentValidation.TestHelper;
 using Xunit;
 
-namespace FleetJourney.Application.Tests.Validation.CarPool;
+namespace FleetJourney.Tests.Validation.CarPool;
 
-public sealed class CarResponseValidatorTests
+public sealed class UpdateCarRequestValidatorTests
 {
-    private readonly CarResponseValidator _validator;
+    private readonly UpdateCarRequestValidator _validator;
 
-    public CarResponseValidatorTests()
+    public UpdateCarRequestValidatorTests()
     {
-        _validator = new CarResponseValidator();
+        _validator = new UpdateCarRequestValidator();
     }
 
     [Theory]
-    [InlineData("")]
     [InlineData(null)]
+    [InlineData("")]
     public async Task SetLicensePlateNumber_WithInvalidLicensePlateNumber_ShouldReturnValidationError(string licensePlateNumber)
     {
         // Arrange
-        var carResponse = new CarResponse
+        var updateCarRequest = new UpdateCarRequest
         {
-            Id = Guid.NewGuid(),
             LicensePlateNumber = licensePlateNumber,
             Brand = "Valid Brand",
             Model = "Valid Model",
             EndOfLifeMileage = 10000,
             MaintenanceInterval = 5000,
-            CurrentMileage = 5000,
-            Trips = null
+            CurrentMileage = 5000
         };
 
         // Act
-        var result = await _validator.TestValidateAsync(carResponse);
+        var result = await _validator.TestValidateAsync(updateCarRequest);
 
         // Assert
-        result.ShouldHaveValidationErrorFor(response => response.LicensePlateNumber)
+        result.ShouldHaveValidationErrorFor(request => request.LicensePlateNumber)
             .WithErrorMessage("License plate number is required.");
     }
 
+    [Fact]
+    public async Task SetLicensePlateNumber_WithInvalidFormat_ShouldReturnValidationError()
+    {
+        // Arrange
+        var updateCarRequest = new UpdateCarRequest
+        {
+            LicensePlateNumber = "InvalidFormat",
+            Brand = "Valid Brand",
+            Model = "Valid Model",
+            EndOfLifeMileage = 10000,
+            MaintenanceInterval = 5000,
+            CurrentMileage = 5000
+        };
+
+        // Act
+        var result = await _validator.TestValidateAsync(updateCarRequest);
+
+        // Assert
+        result.ShouldHaveValidationErrorFor(request => request.LicensePlateNumber)
+            .WithErrorMessage("Invalid license plate number. Format should be 123-AB-45.");
+    }
+
     [Theory]
-    [InlineData("")]
     [InlineData(null)]
+    [InlineData("")]
     public async Task SetBrand_WithInvalidBrand_ShouldReturnValidationError(string brand)
     {
         // Arrange
-        var carResponse = new CarResponse
+        var updateCarRequest = new UpdateCarRequest
         {
-            Id = Guid.NewGuid(),
             LicensePlateNumber = "Valid License Plate Number",
             Brand = brand,
             Model = "Valid Model",
             EndOfLifeMileage = 10000,
             MaintenanceInterval = 5000,
-            CurrentMileage = 5000,
-            Trips = null
+            CurrentMileage = 5000
         };
 
         // Act
-        var result = await _validator.TestValidateAsync(carResponse);
+        var result = await _validator.TestValidateAsync(updateCarRequest);
 
         // Assert
-        result.ShouldHaveValidationErrorFor(response => response.Brand)
+        result.ShouldHaveValidationErrorFor(request => request.Brand)
             .WithErrorMessage("Brand is required.");
     }
 
     [Theory]
-    [InlineData("")]
     [InlineData(null)]
+    [InlineData("")]
     public async Task SetModel_WithInvalidModel_ShouldReturnValidationError(string model)
     {
         // Arrange
-        var carResponse = new CarResponse
+        var updateCarRequest = new UpdateCarRequest
         {
-            Id = Guid.NewGuid(),
             LicensePlateNumber = "Valid License Plate Number",
             Brand = "Valid Brand",
             Model = model,
             EndOfLifeMileage = 10000,
             MaintenanceInterval = 5000,
-            CurrentMileage = 5000,
-            Trips = null
+            CurrentMileage = 5000
         };
 
         // Act
-        var result = await _validator.TestValidateAsync(carResponse);
+        var result = await _validator.TestValidateAsync(updateCarRequest);
 
         // Assert
-        result.ShouldHaveValidationErrorFor(response => response.Model)
+        result.ShouldHaveValidationErrorFor(request => request.Model)
             .WithErrorMessage("Model is required.");
     }
 
@@ -97,23 +113,21 @@ public sealed class CarResponseValidatorTests
     public async Task SetEndOfLifeMileage_WithInvalidEndOfLifeMileage_ShouldReturnValidationError(uint endOfLifeMileage)
     {
         // Arrange
-        var carResponse = new CarResponse
+        var updateCarRequest = new UpdateCarRequest
         {
-            Id = Guid.NewGuid(),
             LicensePlateNumber = "Valid License Plate Number",
             Brand = "Valid Brand",
             Model = "Valid Model",
             EndOfLifeMileage = endOfLifeMileage,
             MaintenanceInterval = 5000,
-            CurrentMileage = 5000,
-            Trips = null
+            CurrentMileage = 5000
         };
 
         // Act
-        var result = await _validator.TestValidateAsync(carResponse);
+        var result = await _validator.TestValidateAsync(updateCarRequest);
 
         // Assert
-        result.ShouldHaveValidationErrorFor(response => response.EndOfLifeMileage)
+        result.ShouldHaveValidationErrorFor(request => request.EndOfLifeMileage)
             .WithErrorMessage("End of life mileage must be greater than zero.");
     }
 
@@ -122,23 +136,21 @@ public sealed class CarResponseValidatorTests
     public async Task SetMaintenanceInterval_WithInvalidMaintenanceInterval_ShouldReturnValidationError(uint maintenanceInterval)
     {
         // Arrange
-        var carResponse = new CarResponse
+        var updateCarRequest = new UpdateCarRequest
         {
-            Id = Guid.NewGuid(),
             LicensePlateNumber = "Valid License Plate Number",
             Brand = "Valid Brand",
             Model = "Valid Model",
             EndOfLifeMileage = 10000,
             MaintenanceInterval = maintenanceInterval,
-            CurrentMileage = 5000,
-            Trips = null
+            CurrentMileage = 5000
         };
 
         // Act
-        var result = await _validator.TestValidateAsync(carResponse);
+        var result = await _validator.TestValidateAsync(updateCarRequest);
 
         // Assert
-        result.ShouldHaveValidationErrorFor(response => response.MaintenanceInterval)
+        result.ShouldHaveValidationErrorFor(request => request.MaintenanceInterval)
             .WithErrorMessage("Maintenance interval must be greater than zero.");
     }
 
@@ -147,47 +159,21 @@ public sealed class CarResponseValidatorTests
     public async Task SetCurrentMileage_WithInvalidCurrentMileage_ShouldReturnValidationError(uint currentMileage)
     {
         // Arrange
-        var carResponse = new CarResponse
+        var updateCarRequest = new UpdateCarRequest
         {
-            Id = Guid.NewGuid(),
             LicensePlateNumber = "Valid License Plate Number",
             Brand = "Valid Brand",
             Model = "Valid Model",
             EndOfLifeMileage = 10000,
             MaintenanceInterval = 5000,
-            CurrentMileage = currentMileage,
-            Trips = null
+            CurrentMileage = currentMileage
         };
 
         // Act
-        var result = await _validator.TestValidateAsync(carResponse);
+        var result = await _validator.TestValidateAsync(updateCarRequest);
 
         // Assert
-        result.ShouldHaveValidationErrorFor(response => response.CurrentMileage)
+        result.ShouldHaveValidationErrorFor(request => request.CurrentMileage)
             .WithErrorMessage("Current mileage must be greater than zero.");
-    }
-
-    [Fact]
-    public async Task SetId_WithEmptyId_ShouldReturnValidationError()
-    {
-        // Arrange
-        var carResponse = new CarResponse
-        {
-            Id = Guid.Empty,
-            LicensePlateNumber = "Valid License Plate Number",
-            Brand = "Valid Brand",
-            Model = "Valid Model",
-            EndOfLifeMileage = 10000,
-            MaintenanceInterval = 5000,
-            CurrentMileage = 5000,
-            Trips = null
-        };
-
-        // Act
-        var result = await _validator.TestValidateAsync(carResponse);
-
-        // Assert
-        result.ShouldHaveValidationErrorFor(response => response.Id)
-            .WithErrorMessage("Id is required.");
     }
 }
