+++ FleetJourney.sln
@@ -19,7 +19,7 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 		docker-compose.yml = docker-compose.yml
 	EndProjectSection
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FleetJourney.Application.Tests", "FleetJourney.Tests\FleetJourney.Application.Tests\FleetJourney.Application.Tests.csproj", "{CCB236C6-38BC-46F0-8B85-7CEA2FCEE31B}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FleetJourney.Tests", "FleetJourney.Tests\FleetJourney.Tests.csproj", "{456BAAEF-86B7-4010-BE8A-B89414D69808}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -55,9 +55,9 @@ Global
 		{A255B27E-B4F6-4B2D-84EE-4EB881D60D26}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{A255B27E-B4F6-4B2D-84EE-4EB881D60D26}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{A255B27E-B4F6-4B2D-84EE-4EB881D60D26}.Release|Any CPU.Build.0 = Release|Any CPU
-		{CCB236C6-38BC-46F0-8B85-7CEA2FCEE31B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{CCB236C6-38BC-46F0-8B85-7CEA2FCEE31B}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{CCB236C6-38BC-46F0-8B85-7CEA2FCEE31B}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{CCB236C6-38BC-46F0-8B85-7CEA2FCEE31B}.Release|Any CPU.Build.0 = Release|Any CPU
+		{456BAAEF-86B7-4010-BE8A-B89414D69808}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{456BAAEF-86B7-4010-BE8A-B89414D69808}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{456BAAEF-86B7-4010-BE8A-B89414D69808}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{456BAAEF-86B7-4010-BE8A-B89414D69808}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 EndGlobal
